Reorganized the files to avoid further conflicts.
