+++ Projects/Project2/network_algo.py
@@ -1,11 +1,5 @@
-import numpy as np
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-import math
-import numpy.linalg as la
-from mpl_toolkits import mplot3d
-from import_data import generate_data, concatenate
-from testing import *
+"""Network class and main algorithm for training the network."""
+from test_model import *
 
 # Add parameters for plotting. 
 mpl.rcParams['figure.titleweight'] = "bold"
@@ -19,6 +13,7 @@ class Parameters:
         self.W_k = np.random.randn(K,d,d)
         self.b_k = np.random.randn(K,d,1)
         self.b_k_I = np.zeros((K,d,I))
+
         for i in range(K):
             self.b_k_I[i,:,:] = self.b_k[i,:,:]
 
@@ -29,10 +24,7 @@ class Parameters:
         # For Adam descent.
         self.beta_1 = 0.9; self.beta_2 = 0.999; self.alpha = 0.01; self.epsilon = 10E-8
         self.v = [0]; self.m = [0]
-        
-    # Perhaps this class should only contain the parameters 
-    # and the update + root functions should be part of the Network?
-    # Perhaps this makes some sense when seeing the algorithm-function after all. 
+     
     def update_parameters(self,gradient,method,tau,j):
         """Update the parameters via Vanilla Gradient Method or Adam Descent."""
         if(method == "vanilla"):
@@ -67,7 +59,6 @@ class Parameters:
             v_j[i] = np.sqrt(v_j[i])
         return v_j
 
-
 class Network:
     """Class for the Neural Network (ResNet)."""
     def __init__(self,K,d,I,h,Z_0,c):
@@ -94,13 +85,11 @@ class Network:
         val = 1-np.tanh(x)**2
         return val
 
-
     def eta(self, x):
         """Hypothesis function. Appears in the final layer of the neural network."""
         val = x
         #val = 0.5*(1+ np.tanh(x/2))
         return val
-
 
     def eta_der(self, x):
         """The derivative of the hypothesis function."""
@@ -118,44 +107,32 @@ class Network:
             
         ZT_K = np.transpose(self.Z_list[-1,:,:]) #Enklere notasjon
         one_vec = np.ones(self.I)
-        '''
-        print("comp1")
-        print(ZT_K @ self.theta.w)
-        print("comp2")
-        print(self.theta.my*one_vec)
-        print("input Z")
-        print(ZT_K @ self.theta.w + self.theta.my*one_vec)
-        '''
+        
         Y = self.eta(ZT_K @ self.theta.w + self.theta.my*one_vec)
         self.Y = Y
 
-    
-        
-    
     def back_propagation(self):
         """Calculate and return the gradient of the objective function."""
         ZT_K = np.transpose(self.Z_list[-1,:,:])
         one_vec = np.ones(self.I)
         J_der_my =  np.transpose(one_vec) @ (self.Y-self.c)
-        #eta_der(ZT_K @ self.theta.w + self.theta.my*one_vec) @ (Y-c)#Blir en skalar  
         J_der_omega = self.Z_list[-1,:,:] @ ((self.Y-self.c) * \
                             self.eta_der(ZT_K @ self.theta.w + self.theta.my*one_vec))
         
         P_K = np.outer(self.theta.w,(self.Y-self.c)*self.eta_der(ZT_K @ \
-                                            self.theta.w + self.theta.my*one_vec)) #Blir en dxI matrise
+                                            self.theta.w + self.theta.my*one_vec)) 
         
         
-        P_list = np.zeros((self.K,self.d,self.I)) #K matriser, skal ikke ha med P_0
-        P_list[-1,:,:] = P_K      #Legger P_K bakerst i P_list
-        for i in range(self.K-1,0,-1):  #Starter på P_k(=indeks K-1) og helt til og med P_1(=indeks 0)
+        P_list = np.zeros((self.K,self.d,self.I)) 
+        P_list[-1,:,:] = P_K 
+        for i in range(self.K-1,0,-1):
             P_list[i-1,:,:] = P_list[i,:,:] + self.h*np.transpose(self.theta.W_k[i-1,:,:]) @ \
             (self.sigma_der(self.theta.W_k[i-1,:,:] @ self.Z_list[i-1,:,:] + \
                             self.theta.b_k_I[i-1,:,:]) * P_list[i,:,:])
 
         J_der_W = np.zeros((self.K,self.d,self.d))
         J_der_b = np.zeros((self.K,self.d,1))
-        one_vec = np.ones((self.I,1))  #Må gjøre vec_I til en matrise av en eller annen grunn
-        #P_Kk går fra P_1(=indeks 0) til P_K(=indeks K-1)
+        one_vec = np.ones((self.I,1)) 
         
         for i in range(self.K):
             val = P_list[i,:,:] * self.sigma_der(self.theta.W_k[i,:,:] @ \
@@ -167,9 +144,6 @@ class Network:
         
         return gradient
     
-    # SGD should be used when calculating the gradient, to reduce the computations.
-    # We should at least test and see how much it affects the convergence. 
-    # This does not draw without replacement however, does it?
     def stochastic_elements(self,Z_0, C, chunk): 
         """Pick out a fixed amount of elements from Z."""        
         start = np.random.randint(self.I-chunk)
@@ -178,7 +152,7 @@ class Network:
         return Z0_chunk, C_chunk 
     
     def embed_test_input(self, test_input, test_output):
-    
+        """Embed the input into d-dimensional space."""
         I_new = test_input.shape[1]
         self.I = I_new
 
@@ -190,12 +164,6 @@ class Network:
         for i in range(self.K):
             self.theta.b_k_I[i,:,:] = self.theta.b_k[i,:,:]
 
-
-
-    # Perhaps there should be one class per part of the Hamiltonian and
-    # this hould be a method of those classes? (or an abstract class 
-    # for the complete Hamiltonian?)
-    # Also fine to leave it here, since it uses all the attributes from the object.  
     def Hamiltonian_gradient(self):
         """Calculate the gradient of F, according to the theoretical derivation.
         
@@ -215,17 +183,14 @@ class Network:
         
         return gradient
 
-    
-
-
-def algorithm(I,d,K,h,iterations, tau, chunk, function,domain,scaling, alpha, beta):
+def algorithm(I, d, d0, K, h, iterations, tau, chunk, function, domain, scaling, alpha, beta):
     """Main training algorithm."""
       
-    input = generate_input(function,domain,d0,I,d)
-    output = get_solution(function,input,d,I,d0)
+    inp = generate_input(function,domain,d0,I,d)
+    output = get_solution(function,inp,d,I,d0)
 
     if scaling:
-        input, a1, b1 = scale_data(alpha,beta,input)
+        inp, a1, b1 = scale_data(alpha,beta,inp)
         output, a2, b2 = scale_data(alpha,beta,output)
 
 
@@ -240,13 +205,12 @@ def algorithm(I,d,K,h,iterations, tau, chunk, function,domain,scaling, alpha, be
     counter = 0
     for j in range(1,iterations+1):
         
-        
         NN.forward_function()
         gradient = NN.back_propagation()
         NN.theta.update_parameters(gradient,"adams",tau,j)
 
         if counter < I/chunk - 1:
-            NN.Z_list[0,:,:] = input[:,chunk*(counter+1):chunk*(counter+2)]
+            NN.Z_list[0,:,:] = inp[:,chunk*(counter+1):chunk*(counter+2)]
             NN.c = output[chunk*(counter + 1):chunk*(counter + 2)]
         else:
             counter = 0
@@ -266,127 +230,3 @@ def algorithm(I,d,K,h,iterations, tau, chunk, function,domain,scaling, alpha, be
     plt.show()
     
     return NN
-
-
-
-""" 
-Below, we train and test the neural network with the provided test functions.
-"""
-
-I = 1000 # Amount of points ran through the network at once. 
-K = 20 # Amount of hidden layers in the network.
-d = 2 # Dimension of the hidden layers in the network. 
-h = 0.05 # Scaling of the activation function application in algorithm.  
-iterations = 2000 #Number of iterations in the Algorithm 
-tau = 0.1 #For the Vanilla Gradient method
-
-#For scaling
-scaling = False
-alpha = 0.2
-beta = 0.8 
-
-#================#
-#Test function 1 #
-#================#
-
-"""
-d0 = 1 # Dimension of the input layer. 
-domain = [-2,2]
-chunk = int(I/10)
-def test_function1(x):
-    return 0.5*x**2
-
-NN = algorithm(I,d,K,h,iterations, tau, chunk, test_function1,domain,scaling, alpha, beta)
-test_input = generate_input(test_function1,domain,d0,I,d)
-
-#The a's and b's are for potential scaling fo the data
-output, a1, b1, a2, b2 = testing(NN, test_input, test_function1, domain, d0, d, I, scaling, alpha, beta)
-plot_graph_and_output(output, test_input, test_function1, domain, d0,d, scaling, alpha, beta, a1, b1, a2, b2)
-"""
-#================#
-#Test function 2 #
-#================#
-"""
-d0 = 1 # Dimensin of the input layer. 
-domain = [-2,2]
-chunk = int(I/10)
-def test_function2(x):
-    return 1 - np.cos(x)
-
-NN = algorithm(I,d,K,h,iterations,tau,chunk, test_function2,domain, scaling, alpha, beta)
-test_input = generate_input(test_function2,domain,d0,I,d)
-
-#The a's and b's are for potential scaling fo the data
-output, a1, b1, a2, b2 = testing(NN, test_input, test_function2, domain, d0, d, I, scaling, alpha, beta)
-plot_graph_and_output(output, test_input, test_function2, domain, d0,d, scaling, alpha, beta, a1, b1, a2, b2)
-"""
-
-#================#
-#Test function 3 #
-#================#
-"""
-d0 = 2
-d = 4
-domain = [[-2,2],[-2,2]]
-chunk = int(I/10)
-def test_function3(x,y):
-    return 0.5*(x**2 + y**2)
-
-NN = algorithm(I,d,K,h,iterations, tau, chunk, test_function3,domain,scaling,alpha,beta)
-test_input = generate_input(test_function3,domain,d0,I,d)
-
-#The a's and b's are for potential scaling fo the data
-output, a1, b1, a2, b2 = testing(NN, test_input, test_function3, domain, d0, d, I, scaling, alpha, beta)
-plot_graph_and_output(output, test_input, test_function3, domain, d0,d, scaling, alpha, beta, a1, b1, a2, b2)
-
-"""
-#================#
-#Test function 4 #
-#================#
-"""
-
-d0 = 2
-d = 4
-domain = [[-2,2],[-2,2]]
-chunk = int(I/10)
-def test_function4(x,y):
-    return -1/np.sqrt(x**2 + y**2)
-
-NN = algorithm(I,d,K,h,iterations, tau, chunk, test_function4,domain,scaling,alpha,beta)
-test_input = generate_input(test_function4,domain,d0,I,d)
-
-#The a's and b's are for potential scaling fo the data
-output, a1, b1, a2, b2 = testing(NN, test_input, test_function4, domain, d0, d, I, scaling, alpha, beta)
-plot_graph_and_output(output, test_input, test_function4, domain, d0,d, scaling, alpha, beta, a1, b1, a2, b2)
-
-"""
-
-## Test the Hamiltonian function below!
-# Test with Kepler two-body problem.
-
-def T(p1,p2):
-    return 0.5*(p1**2 + p2**2)
-
-def exact_grad_T(p):
-
-    return np.array([p[0],p[1]])
-
-# Make neural network for T.
-d0 = 2
-d = 4
-domain = [[-2,2],[-2,2]]
-I = 100
-K = 10
-iterations = 1000
-chunk = int(I/10)
-
-NNT = algorithm(I,d,K,h,iterations, tau, chunk,T,domain,scaling,alpha,beta)
-
-test_input = generate_input(exact_grad_T, domain, d0, I, d)
-output, a1, b1, a2, b2 = testing(NNT,test_input,T,domain,d0,d,I,scaling,alpha,beta)
-grad = NNT.Hamiltonian_gradient()
-grad_scaled = grad[:d0,:]
-#print(grad_scaled)
-#print(exact_grad_T(test_input))
-print(la.norm(grad[:d0,:] - exact_grad_T(test_input[:d0,:]))) # Her har de forskjellig dimensjon...
-
