+++ Projects/Project2/test_given_data.py
@@ -1,0 +1,6 @@
+"""Train and test networks on the given data (from unknown Hamiltonian function)."""
+from network_algo import *
+from import_data import generate_data, concatenate
+
+# Run training and tests on the given data here! (from import_data)
+# Need to finish the numerical methods first I think!
